Add WeatherApiService with Get method
A new service class, WeatherApiService, has been created. This class makes use of HttpClient and includes a Get method to fetch weather data from an API. This data is then deserialized into a Root object and returned. Currently, GetAll and Get(int id) methods throw an exception as they are yet to be implemented.