Refactor HTML layout in GitHubProjectBox view
This commit updates the HTML markup in the _GitHubProjectBox.cshtml view. The image alt text was added for the project image and a "github-project-content" div was introduced for better structure and readability of the layout. The changes aim to enhance accessibility and improve the structural organization of the project box.