Add icon dictionary to Parameters class
A new static read-only dictionary named "Icons" has been added to the Parameters class. The dictionary contains a mapping of technology names to their respective skill icon URLs which are hosted on an external CDN. This will be used to display the respective icons for various technologies on the website.