Update project descriptions and images in DataInitializer
In this commit, the project descriptions in the DataInitializer class have been updated to provide more layout for each project. Additionally, the project images' URLs have been updated to point to a .webp version instead of a .png for improved loading speed. A new project 'RESTful API' has also been added to the list of projects.