Refactor GithubApiService and WeatherApiService implementation
Changes have been made to the GithubApiService and WeatherApiService to implement the interfaces IGithubProjectService and IWeatherService respectively rather than IHttpService. The 'Get' methods have been updated to reflect the new data to be fetched. In addition, the Root model in the ViewModel has been slimmed down to only contain properties necessary for the current application requirements.