Refactored _GithubProjectBox.cshtml view
This commit includes refactoring the _GithubProjectBox.cshtml view by updating layout structure and case consistency. We have removed hardcoded styles from the title tag and merely use the defined classes. Project and Demo URLs are now capitalized consistently and the LiveDemoUrl section only renders if the URL is not null or empty.